Add files via upload
Description
This repository contains the work done for Task 1 of my internship at VilearnX Advanced Technology. The primary objective of this task is to perform Exploratory Data Analysis (EDA) on a chosen dataset to uncover its underlying structure, distributions, correlations, and outliers. Through a series of data analysis and visualization techniques, we aim to gain valuable insights and a deeper understanding of the dataset's characteristics.